refactor(trusty-search): split lazy_loader.rs under 500-line cap (closes line-cap CI)
lazy_loader.rs (606 lines, new file, not allowlisted) violates the 500-line hard
cap enforced by scripts/check_line_cap.sh. Split into a focused 4-file module:

  - lazy_loader/env.rs    — LAST_QUERIED_WRITE_INTERVAL_SECS, warmboot_max_indexes(),
                            cold_reload_timeout()
  - lazy_loader/store.rs  — select_warmboot_entries(), ColdIndexStore
  - lazy_loader/loader.rs — get_or_load_index(), LazyLoadError
  - lazy_loader/mod.rs    — re-export facade + #[cfg(test)] block

All public symbols re-exported from mod.rs so callers (start.rs, server/state.rs,
server/search.rs) need no import changes. All tests pass; clippy clean; fmt clean.
